requirements.txt
APScheduler
black
contextily
ExifRead
Flask
Flask-Bcrypt
Flask-Login
flask-mongoengine
folium
fsspec
geopandas
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
gcsfs
jupyter
isort
mypy
flake8 
pyshp
torchserve
types-requests
yolov5>=5.0.7